feat: add global exception handling

+++ ems/src/main/java/com/ems/service/EmployeeService.java
@@ -1,6 +1,7 @@
 package com.ems.service;
 
 import com.ems.entity.Employee;
+import com.ems.exception.EmployeeNotFoundException;
 import com.ems.repository.EmployeeRepository;
 import org.springframework.stereotype.Service;
 
@@ -15,19 +16,32 @@ public class EmployeeService {
         this.employeeRepository = employeeRepository;
     }
 
+    // Add Employee
     public Employee addEmployee(Employee employee) {
         return employeeRepository.save(employee);
     }
 
+    // Get All Employees
     public List<Employee> getAllEmployees() {
         return employeeRepository.findAll();
     }
 
+    // Get Employee by ID
     public Employee getEmployeeById(Long id) {
-        return employeeRepository.findById(id).orElse(null);
+        return employeeRepository.findById(id)
+                .orElseThrow(() ->
+                        new EmployeeNotFoundException(
+                                "Employee not found with id: " + id));
     }
 
+    // Delete Employee
     public void deleteEmployee(Long id) {
+
+        if (!employeeRepository.existsById(id)) {
+            throw new EmployeeNotFoundException(
+                    "Employee not found with id: " + id);
+        }
+
         employeeRepository.deleteById(id);
     }
 }
